Refactor signer provider and profile config loader to base interface class and implementations
